move metrics from train.py to separate file
